+++ examples/cttc-nr-demo.cc
@@ -49,7 +49,7 @@ main (int argc, char *argv[])
   bool udpFullBuffer = false;
   int32_t fixedMcs = -1;
   uint16_t gNbNum = 1;
-  uint16_t ueNumPergNb = 1;
+  uint16_t ueNumPergNb = 0;
   bool cellScan = false;
   double beamSearchAngleStep = 10.0;
   uint16_t numerologyBwp1 = 4;
@@ -215,6 +215,8 @@ main (int argc, char *argv[])
 
   mmWaveHelper->SetGnbAntennaAttribute ("NumRows", UintegerValue (4));
   mmWaveHelper->SetGnbAntennaAttribute ("NumColumns", UintegerValue (8));
+
+  mmWaveHelper->SetGnbPhyAttribute ("Numerology", UintegerValue (2));
 
   // install mmWave net devices
   NetDeviceContainer enbNetDev = mmWaveHelper->InstallGnbDevice (gridScenario.GetBaseStations (), allBwps);