+++ model/mmwave-mac-sched-sap.h
@@ -172,6 +172,10 @@ public:
   virtual uint16_t GetBwpId () const = 0;
 
   virtual uint16_t GetCellId () const = 0;
+
+  virtual uint32_t GetSymbolsPerSlot () const = 0;
+
+  virtual Time GetSlotPeriod () const = 0;
 };
 
 std::ostream & operator<< (std::ostream & os, MmWaveMacSchedSapProvider::SchedDlRlcBufferReqParameters const & p);